+++ org.eclipse.buildship.oomph.edit/build.properties
@@ -11,5 +11,5 @@ bin.includes = .,\
                plugin.xml,\
                plugin.properties
 jars.compile.order = .
-source.. = src/main/java/
+source.. = src/main/java-generated/
 output.. = bin/
