+++ org.eclipse.buildship.oomph/build.properties
@@ -11,6 +11,6 @@ bin.includes = .,\
                plugin.xml,\
                plugin.properties
 jars.compile.order = .
-source.. = src/main/java/,\
+source.. = src/main/java-generated/,\
            src/main/resources/
 output.. = bin/
